dataset: new modularize, no verify
