+++ Cargo.toml
@@ -10,6 +10,7 @@ bio = "1.1.0"
 chrono = {version = "0.4.31", features = ["serde"] }
 clap = { version = "4.2.7", features = ["derive"] }
 color-eyre = "0.6.2"
+either = "1.9.0"
 env_logger = "0.10.0"
 image = "0.24.6"
 indicatif = {version = "0.17.5", features = ["rayon"]}
